Unit tests passed
The HallMonitor passed the challenge's unit tests, and thus I am dubbing it a massive success with no errors!

Really, there's no input validation going on, so that would be something to address in a future update. Check to make sure that the non-H Objects are integers (as that's the stand-in for rooms), and return false if they aren't integers.