EqualityOfValues final commit
I made equal() as lean as possible: it returns a nested ternary operator value. It's not the easiest thing to read, but by golly is it concise. Funnily, no other responses on the solution page are anywhere near as short, but obviously they're almost all much easier to parse/read at a glance. That was fun!