Updated SevenBoom Notes
I've filled out the PROTOTYPE section of my notes. In an attempt to increase the speed of the function on single-digit integers, I've got an if < 10 check going, where if it's true, it's an int to int comparison, but if false the input integer is converted to a string and each element is compared to "7". Time to write and test it!