PositiveDominant Debugging
Good news, everyone! My PositiveDominant.java ran with no errors and three failures. The failures were that the boolean was returning as the opposite value that it should have, so I had to crack open the debugger and step through the execution to figure out what was going on. I was trying to account for there being at least one unique positive integer, so my || m == 0 was running with each int l loop, which was artificially inflating the sumUniquePositive value. I need to do some kind of validation earlier to prevent the insertion of duplicate values in different array indices that have to be double-checked. Screenshot of debugging to follow.