PositiveDominant final commit
I finally implemented PositiveDominant using solely native int[] objects, and this commit adds them to this repo. I didn't struggle at all with how to implement it using an ArrayList<Integer>, but it took me a significantly longer time to figure it out with int[]. Please read the updated PositiveDominantNotes, the Post-Activity section within, and the // comments in each of the classes to follow along with my thought process!