TimeConverter tested/working commit
I've completed the TimeConverter to pass the original prompt's unit tests, with a few additions of my own. Tests 1-4 were all included with the prompt, and 5-7 were my addition. I also thoroughly commented TimeConverter.java I also filled out the Post-Activity section of the notes. Looks to me like it works for 12- to 24-hour and 24- to 12-hour conversions.